refactor,update: update dependencies, add logging and null checks, move SubjectCertificatePolicies to certifcate package
