+++ src/test/java/eu/webeid/security/validator/ocsp/OcspServiceProviderTest.java
@@ -82,5 +82,4 @@ class OcspServiceProviderTest {
                 service2018.validateResponderCertificate(wrongResponderCert, new Date(1630000000000L)));
     }
 
-
 }
